+++ YourName/YourName/Sources/Data Layer/Network/MeetuNetwork/NetworkService.swift
@@ -21,7 +21,6 @@ final class NetworkService: NetworkServing {
         return self.provider.rx.request(endpoint)
             .asObservable()
             .map(MeetuResponse<API.Response>.self)
-            .do(onNext: { print("📝 response -> \($0)") })
             .map { $0.data }
             .filterNil()
     }
